+++ src/BitcoinGateway.php
@@ -1,0 +1,49 @@
+<?php
+
+/**
+ * Title: WooCommerce Bitcoin gateway
+ * Description:
+ * Copyright: Copyright (c) 2005 - 2016
+ * Company: Pronamic
+ *
+ * @author Remco Tolsma
+ * @version unreleased
+ * @since unreleased
+ */
+class Pronamic_WP_Pay_Extensions_WooCommerce_BitcoinGateway extends Pronamic_WP_Pay_Extensions_WooCommerce_Gateway {
+	/**
+	 * The unique ID of this payment gateway
+	 *
+	 * @var string
+	 */
+	const ID = 'pronamic_pay_bitcoin';
+
+	//////////////////////////////////////////////////
+
+	/**
+	 * Constructs and initialize an bank transfer gateway
+	 */
+	public function __construct() {
+		$this->id           = self::ID;
+		$this->method_title = __( 'Bitcoin', 'pronamic_ideal' );
+		$this->payment_method = Pronamic_WP_Pay_PaymentMethods::BITCOIN;
+
+		// @see https://github.com/woothemes/woocommerce/blob/v1.6.6/classes/gateways/class-wc-payment-gateway.php#L24
+		$this->has_fields = false;
+
+		parent::__construct();
+	}
+
+	//////////////////////////////////////////////////
+
+	/**
+	 * Initialise form fields
+	 */
+	function init_form_fields() {
+		parent::init_form_fields();
+
+		$this->form_fields['enabled']['label']       = __( 'Enable Bitcoin', 'pronamic_ideal' );
+		$this->form_fields['description']['default'] = '';
+		$this->form_fields['icon']['default']        = '';
+	}
+}